feat: recipe creation using from-or-html endpoint - extract advanced mode
