+++ nodes/Mealie/resources/RecipeCrudResource.ts
@@ -1,5 +1,5 @@
 import { INodeProperties, INodePropertyOptions } from "n8n-workflow";
-import { CreateRecipeFromHtmlOrJsonOperation } from "./recipeCrud";
+import { CreateRecipe, CreateRecipeFromHtmlOrJsonOperation } from "./recipeCrud";
 import { MealieN8nResource } from "../generic/MealieN8nResource";
 
 export class RecipeCrudResource implements MealieN8nResource {
@@ -20,10 +20,12 @@ export class RecipeCrudResource implements MealieN8nResource {
         },
       },
       options: [
+        CreateRecipe.Operation,
         CreateRecipeFromHtmlOrJsonOperation.Operation,
       ],
       default: '',
     },
+    ...CreateRecipe.Fields,
     ...CreateRecipeFromHtmlOrJsonOperation.Fields,
   ];
 }
