+++ nodes/Mealie/resources/recipeCrud/CreateRecipeFromHtmlOrJsonOperation.ts
@@ -4,7 +4,7 @@ import { MealieN8nOperation } from '../../generic/MealieN8nOperation';
 export class CreateRecipeFromHtmlOrJsonOperation implements MealieN8nOperation {
 	static readonly OperationId = 'createRecipeFromHtmlOrJson';
 	static readonly Operation: INodePropertyOptions = {
-		name: 'Create Recipe from HTML or JSON',
+		name: 'Create Recipe From HTML or JSON',
 		value: CreateRecipeFromHtmlOrJsonOperation.OperationId,
 		action: 'Create a recipe from HTML or JSON-LD schema',
 		description: 'Takes in raw HTML or a https://schema.org/Recipe object as a JSON string and parses it like a URL',
@@ -58,8 +58,6 @@ export class CreateRecipeFromHtmlOrJsonOperation implements MealieN8nOperation {
 			send: {
 				preSend: [
 					async function (this, requestOptions) {
-						const advancedMode = this.getNodeParameter('advancedMode') as boolean;
-						
 						// Get additional options
 						const additionalOptions = this.getNodeParameter('additionalOptions', {}) as {
 							includeTags?: boolean;
@@ -68,108 +66,8 @@ export class CreateRecipeFromHtmlOrJsonOperation implements MealieN8nOperation {
 						const includeTags = additionalOptions.includeTags ?? true;
 						const url = additionalOptions.url ?? '';
 
-						let data: string;
-
-						if (advancedMode) {
-							// Raw JSON mode - use the data directly
-							data = this.getNodeParameter('rawJsonData') as string;
-						} else {
-							// Simple mode - build JSON-LD from individual fields
-							const recipeName = this.getNodeParameter('recipeName') as string;
-							const description = this.getNodeParameter('description', '') as string;
-							const recipeCategory = this.getNodeParameter('recipeCategory', '') as string;
-							const recipeCuisine = this.getNodeParameter('recipeCuisine', '') as string;
-							const keywords = this.getNodeParameter('keywords', '') as string;
-							const recipeYield = this.getNodeParameter('recipeYield', '') as string;
-							const prepTime = this.getNodeParameter('prepTime', '') as string;
-							const cookTime = this.getNodeParameter('cookTime', '') as string;
-							const totalTime = this.getNodeParameter('totalTime', '') as string;
-							const imageUrl = this.getNodeParameter('imageUrl', '') as string;
-
-							// Check if using advanced list mode for ingredients
-							const ingredientsMode = this.getNodeParameter('ingredientsMode') as string;
-							let ingredients: string[] = [];
-
-							if (ingredientsMode === 'json') {
-								// JSON array mode
-								const ingredientsJson = this.getNodeParameter('ingredientsJson') as string;
-								try {
-									ingredients = JSON.parse(ingredientsJson);
-									if (!Array.isArray(ingredients)) {
-										throw new Error('Ingredients JSON must be an array');
-									}
-								} catch (error) {
-									throw new Error(`Invalid JSON for ingredients: ${error.message}`);
-								}
-							} else {
-								// UI mode (fixedCollection)
-								const ingredientsData = this.getNodeParameter('ingredientsUi', {}) as {
-									ingredientValues?: Array<{ ingredient: string }>;
-								};
-								ingredients = ingredientsData.ingredientValues?.map(item => item.ingredient) || [];
-							}
-
-							// Check if using advanced list mode for instructions
-							const instructionsMode = this.getNodeParameter('instructionsMode') as string;
-							let instructions: Array<{ '@type': string; text: string }> = [];
-
-							if (instructionsMode === 'json') {
-								// JSON array mode
-								const instructionsJson = this.getNodeParameter('instructionsJson') as string;
-								try {
-									const instructionsArray = JSON.parse(instructionsJson);
-									if (!Array.isArray(instructionsArray)) {
-										throw new Error('Instructions JSON must be an array');
-									}
-									instructions = instructionsArray.map(step => ({
-										'@type': 'HowToStep',
-										text: step,
-									}));
-								} catch (error) {
-									throw new Error(`Invalid JSON for instructions: ${error.message}`);
-								}
-							} else {
-								// UI mode (fixedCollection)
-								const instructionsData = this.getNodeParameter('instructionsUi', {}) as {
-									instructionValues?: Array<{ step: string }>;
-								};
-								instructions = instructionsData.instructionValues?.map(item => ({
-									'@type': 'HowToStep',
-									text: item.step,
-								})) || [];
-							}
-
-							// Validate required fields
-							if (ingredients.length === 0) {
-								throw new Error('At least one ingredient is required to create a recipe');
-							}
-
-							if (instructions.length === 0) {
-								throw new Error('At least one instruction step is required to create a recipe');
-							}
-
-							// Build JSON-LD object
-							const jsonLd: Record<string, unknown> = {
-								'@context': 'https://schema.org',
-								'@type': 'Recipe',
-								name: recipeName,
-							};
-
-							// Add optional fields only if they have values
-							if (description) jsonLd.description = description;
-							if (recipeCategory) jsonLd.recipeCategory = recipeCategory;
-							if (recipeCuisine) jsonLd.recipeCuisine = recipeCuisine;
-							if (keywords) jsonLd.keywords = keywords;
-							if (recipeYield) jsonLd.recipeYield = recipeYield;
-							if (prepTime) jsonLd.prepTime = prepTime;
-							if (cookTime) jsonLd.cookTime = cookTime;
-							if (totalTime) jsonLd.totalTime = totalTime;
-							if (imageUrl) jsonLd.image = imageUrl;
-							if (ingredients.length > 0) jsonLd.recipeIngredient = ingredients;
-							if (instructions.length > 0) jsonLd.recipeInstructions = instructions;
-
-							data = JSON.stringify(jsonLd);
-						}
+						// Raw JSON mode - use the data directly
+						const data = this.getNodeParameter('rawJsonData') as string;
 
 						// Build the request body
 						const body: Record<string, unknown> = {
@@ -190,335 +88,6 @@ export class CreateRecipeFromHtmlOrJsonOperation implements MealieN8nOperation {
 	};
 
 	static readonly Fields: INodeProperties[] = [
-		// Advanced Mode Toggle
-		{
-			displayName: 'Advanced Mode',
-			name: 'advancedMode',
-			type: 'boolean',
-			default: false,
-			description: 'Whether to use raw JSON input instead of individual fields',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-				},
-			},
-		},
-
-		// ============ SIMPLE MODE FIELDS ============
-
-		{
-			displayName: 'Recipe Name',
-			name: 'recipeName',
-			type: 'string',
-			required: true,
-			default: '',
-			description: 'The name of the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Description',
-			name: 'description',
-			type: 'string',
-			typeOptions: {
-				rows: 4,
-			},
-			default: '',
-			description: 'A description of the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-
-		// ============ INGREDIENTS SECTION ============
-
-		{
-			displayName: 'Ingredients Input Mode',
-			name: 'ingredientsMode',
-			type: 'options',
-			options: [
-				{
-					name: 'Add One by One',
-					value: 'ui',
-					description: 'Add ingredients individually using the UI',
-				},
-				{
-					name: 'JSON Array',
-					value: 'json',
-					description: 'Provide ingredients as a JSON array (ideal for automation)',
-				},
-			],
-			default: 'ui',
-			description: 'Choose how to input ingredients',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Ingredients',
-			name: 'ingredientsUi',
-			type: 'fixedCollection',
-			typeOptions: {
-				multipleValues: true,
-				sortable: true,
-			},
-			default: {},
-			placeholder: 'Add Ingredient',
-			description: 'List of ingredients for the recipe. At least one ingredient is required.',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-					ingredientsMode: ['ui'],
-				},
-			},
-			options: [
-				{
-					name: 'ingredientValues',
-					displayName: 'Ingredient',
-					values: [
-						{
-							displayName: 'Ingredient',
-							name: 'ingredient',
-							type: 'string',
-							default: '',
-							placeholder: 'e.g., 2 cups flour',
-							description: 'An ingredient with quantity and unit',
-						},
-					],
-				},
-			],
-		},
-		{
-			displayName: 'Ingredients (JSON Array)',
-			name: 'ingredientsJson',
-			type: 'string',
-			typeOptions: {
-				rows: 6,
-			},
-			required: true,
-			default: '["2 cups flour", "1 tsp salt", "1 cup water"]',
-			placeholder: '["ingredient 1", "ingredient 2", "ingredient 3"]',
-			description: 'Ingredients as a JSON array of strings. At least one ingredient is required.',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-					ingredientsMode: ['json'],
-				},
-			},
-		},
-
-		// ============ INSTRUCTIONS SECTION ============
-
-		{
-			displayName: 'Instructions Input Mode',
-			name: 'instructionsMode',
-			type: 'options',
-			options: [
-				{
-					name: 'Add One by One',
-					value: 'ui',
-					description: 'Add instruction steps individually using the UI',
-				},
-				{
-					name: 'JSON Array',
-					value: 'json',
-					description: 'Provide instructions as a JSON array (ideal for automation)',
-				},
-			],
-			default: 'ui',
-			description: 'Choose how to input instructions',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Instructions',
-			name: 'instructionsUi',
-			type: 'fixedCollection',
-			typeOptions: {
-				multipleValues: true,
-				sortable: true,
-			},
-			default: {},
-			placeholder: 'Add Step',
-			description: 'Step-by-step instructions for the recipe. At least one step is required.',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-					instructionsMode: ['ui'],
-				},
-			},
-			options: [
-				{
-					name: 'instructionValues',
-					displayName: 'Step',
-					values: [
-						{
-							displayName: 'Step',
-							name: 'step',
-							type: 'string',
-							typeOptions: {
-								rows: 2,
-							},
-							default: '',
-							placeholder: 'e.g., Preheat the oven to 350°F',
-							description: 'A single instruction step',
-						},
-					],
-				},
-			],
-		},
-		{
-			displayName: 'Instructions (JSON Array)',
-			name: 'instructionsJson',
-			type: 'string',
-			typeOptions: {
-				rows: 8,
-			},
-			required: true,
-			default: '["Preheat the oven to 350°F", "Mix dry ingredients", "Add wet ingredients and stir", "Bake for 30 minutes"]',
-			placeholder: '["step 1", "step 2", "step 3"]',
-			description: 'Instructions as a JSON array of strings. At least one step is required.',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-					instructionsMode: ['json'],
-				},
-			},
-		},
-
-		// ============ OTHER RECIPE FIELDS ============
-
-		{
-			displayName: 'Category',
-			name: 'recipeCategory',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., Dessert, Main Course, Appetizer',
-			description: 'The category of the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Cuisine',
-			name: 'recipeCuisine',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., Italian, Mexican, Japanese',
-			description: 'The cuisine type of the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Keywords',
-			name: 'keywords',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., quick, easy, vegetarian',
-			description: 'Comma-separated keywords for the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Yield',
-			name: 'recipeYield',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., 4 servings, 1 loaf',
-			description: 'The quantity produced by the recipe',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Prep Time',
-			name: 'prepTime',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., PT15M (15 minutes)',
-			description: 'Preparation time in ISO 8601 duration format (PT#H#M)',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Cook Time',
-			name: 'cookTime',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., PT1H (1 hour)',
-			description: 'Cooking time in ISO 8601 duration format (PT#H#M)',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Total Time',
-			name: 'totalTime',
-			type: 'string',
-			default: '',
-			placeholder: 'e.g., PT1H15M (1 hour 15 minutes)',
-			description: 'Total time in ISO 8601 duration format (PT#H#M)',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-		{
-			displayName: 'Image URL',
-			name: 'imageUrl',
-			type: 'string',
-			default: '',
-			placeholder: 'https://example.com/recipe-image.jpg',
-			description: 'URL of the recipe image',
-			displayOptions: {
-				show: {
-					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [false],
-				},
-			},
-		},
-
 		// ============ ADVANCED MODE FIELD ============
 
 		{
@@ -535,7 +104,6 @@ export class CreateRecipeFromHtmlOrJsonOperation implements MealieN8nOperation {
 			displayOptions: {
 				show: {
 					operation: [CreateRecipeFromHtmlOrJsonOperation.OperationId],
-					advancedMode: [true],
 				},
 			},
 		},
