+++ nodes/Mealie/resources/recipeCrud/index.ts
@@ -1,1 +1,2 @@
+export * from './CreateRecipe';
 export * from './CreateRecipeFromHtmlOrJsonOperation';
